requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
psutil==5.9.8
aiohttp==3.9.5
gitpython==3.1.43
